Fixed indentation / tab space bugs
